backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
flask-jwt-extended==4.6.0

python-dotenv==1.0.0
requests==2.32.5

# ✅ OFFICIAL & SUPPORTED GEMINI SDK
google-genai==1.60.0

# Utilities
pydantic==2.12.5
Werkzeug==3.0.1
Flask-Mail==0.9.1
email-validator==2.2.0
gunicorn==21.2.0
